// TypeScipt_Linker/src/enums/enums.ts
/**
 * This enum holds the string representaion of the code elements.
 */
export enum searchStringPartType{
    Namespace = 'Namespace',
    Interface = 'Interface',
    Class = 'Class',
    Method = 'Method',
    Property = 'Property',
    Argument = 'Argument',
    Enum = 'Enum',
  }

  /**
 * This enum holds the string representaion of the code elements.
 */
export enum treeNodeType{
  Namespace = 'Namespace',
  Interface = 'Interface',
  Class = 'Class',
  Method = 'Method',
  Property = 'Property',
  Argument = 'Argument',
  File = 'File',
  Enum = 'Enum',
}